refactor: split storage core helpers
